src/ml-model-management/ml-model-management.service.spec.ts
import { Test, TestingModule } from '@nestjs/testing';
import { MlModelManagementService } from './ml-model-management.service';
import { ConfigService } from '@nestjs/config';
import { createClient } from '@supabase/supabase-js';
import * as tf from '@tensorflow/tfjs-node';

jest.mock('@supabase/supabase-js');
jest.mock('@tensorflow/tfjs-node');

describe('MlModelManagementService', () => {
  let service: MlModelManagementService;
  let mockConfigService: Partial<ConfigService>;
  let mockSupabaseClient: any;

  beforeEach(async () => {
    mockConfigService = {
      get: jest.fn().mockImplementation((key: string) => {
        const config = {
          SUPABASE_URL: 'https://example.supabase.co',
          SUPABASE_KEY: 'mock-key',
          HUGGINGFACE_API_KEY: 'mock-huggingface-key',
          HUGGINGFACE_INFERENCE_ENDPOINT: 'https://api-inference.huggingface.co/models/mock-model',
        };
        return config[key];
      }),
    };

    mockSupabaseClient = {
      from: jest.fn().mockReturnThis(),
      select: jest.fn().mockReturnThis(),
      eq: jest.fn().mockReturnThis(),
      order: jest.fn().mockReturnThis(),
      limit: jest.fn().mockResolvedValue({ data: [{ id: 1, name: 'test-model', version: '1.0', data: {} }], error: null }),
      insert: jest.fn().mockResolvedValue({ error: null }),
      update: jest.fn().mockResolvedValue({ error: null }),
    };

    (createClient as jest.Mock).mockReturnValue(mockSupabaseClient);

    const module: TestingModule = await Test.createTestingModule({
      providers: [
        MlModelManagementService,
        { provide: ConfigService, useValue: mockConfigService },
      ],
    }).compile();

    service = module.get<MlModelManagementService>(MlModelManagementService);
  });

  it('should be defined', () => {
    expect(service).toBeDefined();
  });

  describe('getModel', () => {
    it('should return a model', async () => {
      const result = await service.getModel('test-model');
      expect(result).toEqual({ id: 1, name: 'test-model', version: '1.0', data: {} });
    });
  });

  describe('updateModel', () => {
    it('should update a model', async () => {
      await expect(service.updateModel('test-model', { data: 'new-data' }, { metadata: 'test' })).resolves.not.toThrow();
    });
  });

  describe('listModels', () => {
    it('should list models', async () => {
      mockSupabaseClient.select.mockReturnValueOnce({
        order: jest.fn().mockResolvedValue({ data: [{ name: 'model1' }, { name: 'model2' }], error: null }),
      });

      const result = await service.listModels();
      expect(result).toEqual([{ name: 'model1' }, { name: 'model2' }]);
    });
  });

  describe('runInference', () => {
    it('should run inference', async () => {
      jest.spyOn(global, 'fetch').mockResolvedValueOnce({
        json: jest.fn().mockResolvedValueOnce({ result: 'mocked inference result' }),
      } as any);

      const result = await service.runInference('test-model', { input: 'test input' });
      expect(result).toEqual({ result: 'mocked inference result' });
    });
  });

  // Add more tests for other methods...
});
